It seem to be finished... I mean all- categories and authors fragments, their pagers, searching through them and saving their state and returning to standart (menu) mode... There are some minor bugs (i.e. not hiding actionBars menu buttons on orientation changes and not saving selected author/category while searching and screen rotation) but they are really minor and second caused by hardcoded SearchView by Google (in my non professional opinion).
And we also find and fix stupid bug! I mean crushing on NPE on loading
from bottom (author). I think that was caused by refactoring code not
with help of "find and replace" on selected lines.

So, next we'll check and do telephone mode for all this s**t and,
finally, will start coding FragmentArticle! Hurray!