May be we fix here error that throws while loading from bottom and have publishing lag... We try to fix it by simply deleting all categories/authors rows and start loading from top then;
May be we do not fix it. But after some launches with this bug we
successfully load arts both from top (initial) and then from bottom. 

We need to test it. But we need to simulate this situation on our own
server... And I'm to lazy to do it write now; =)