It's a f*cking hardcore (-code)!
There were a huge problem - in some cases for unknown reason scrollView
scrolls itself while we select it through pager. We f*ck it and start to
rewrite whole fragment (Article fragment) via RecyclerView.

But we faced same promblem here too! We, firstly, think about, that
reason in setting image height at runtime, on it's loading. But after
some awful days we understand, that the reason (in case of RecyclerView)
was in automatic scrolling to the middle of one of its item...

Then we start some research and find, that there is method in
LinearLayoutManager, that called on scroll event. It's called when we
select some Article fragment in some cases when, i.e., Recyclers items
height was bigger then device height. So, after some hours we find an
answer! 

What we do:

1) We add boolean firstScroll to LinearLayoutManager to stop each first
scroll;

2) In scrollVerticallyBy (int dy, RecyclerView.Recycler recycler,
RecyclerView.State state) we check this boolean and do not scroll
(return zero from this method) if it's true and set it to false;

3) We add firing Intent to selected fragment in PagerListener;

4) In onReceive method of fragment we set our firstScroll var to true.

Summary: so when we selected some fragment, it can start to scroll
itself to the middle of currently showing item, but he can't do it! And
each next scrolling event processing as it have to by default
implementation.

There is only one "feature" in it. If fragment don't try to scroll
itself when it's selected, our first try to scroll it manually fails...
But, I think it's acceptable cost for resolving the whole problem. =)